Fix shortcut key for Chinese UI
English UI the save, copy, cut, paste, Insert Table buttons' tips when mouse over the button
are apple command or windows ctrl and shortkey, like this 'save(⌘S)',but in Chinese UI save button's tip is '保存(S)(⌘S)'
the '(S)' part is unnecessary. And the short key for copy and paste should be '复制(⌘C) 粘贴(⌘V)' instead of '复制(Y)(⌘C) 粘贴(P)(⌘V)'
Although the '⌘' character is not displayed corrected, I suggest to  remove the unnecessary part.

Change-Id: Ia565efe573285a255558d5f32f3ed1b8242a8f63